+++ spec/integration/sequence_and_naming_i3s.rb
@@ -1,4 +1,6 @@
 I3S_DEPENDENCIES = {
+  ArtifactBundle: [:DeploymentGroup],
+  DeploymentGroup: [],
   BuildPlan: [:PlanScript],
   DeploymentPlan: [:BuildPlan, :GoldenImage],
   GoldenImage: [:BuildPlan, :OSVolume],
@@ -29,6 +31,10 @@ end
 PLAN_SCRIPT1_NAME = 'Plan_Script_1'.freeze
 PLAN_SCRIPT2_NAME = 'Plan_Script_2'.freeze
 PLAN_SCRIPT1_NAME_UPDATE = 'Plan_Script_1_Updated'.freeze
+
+# Artifact Bundle
+ARTIFACT_BUNDLE1_NAME = 'Artifact_Bundle_1'.freeze
+ARTIFACT_BUNDLE2_NAME = 'Artifact_Bundle_ReadOnly'.freeze
 
 # Build Plan
 BUILD_PLAN1_NAME = 'Build_Plan_1'.freeze