+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 #### New Resources:
 Added full support to Image Streamer Rest API version 300:
+   - Artifact Bundle
    - Build Plan
+   - Deployment Group
    - Deployment Plan
    - Golden Image
    - OS Volume